requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pypdf>=4.0.0
python-docx>=1.1.0
psycopg2-binary>=2.9.9
mcp>=1.0.0
flask>=3.0.0
werkzeug>=3.0.0
